Fix MSVC build compatibility and update build docs
- Fix OpenMP loops to use signed int (MSVC OpenMP 2.0 requirement)
- Rename M_E to M_ELECTRON to avoid cmath macro conflict
- Add missing includes (ParticleManager.h, QApplication, Constants.h)
- Update OCCT libs for 7.9+ (TKSTEP -> TKDESTEP/TKDE)
- Remove Eigen3 version pin (supports both 3.x and 5.x)
- Add Conda-based dependency install instructions for Windows
- Update main README with Cpp3D module description

+++ Cpp3D/src/core/BorisPusher3D.cpp
@@ -14,7 +14,7 @@ void BorisPusher3D::push(ParticleArray& particles, const Grid3D& grid,
     if (np == 0) return;
 
     #pragma omp parallel for schedule(static)
-    for (size_t i = 0; i < np; i++) {
+    for (int i = 0; i < static_cast<int>(np); i++) {
         if (!particles.alive[i]) continue;
 
         double px = particles.x[i];
